+++ packages/smarthr-ui/src/components/DefinitionList/DefinitionListItem.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import {
   type ComponentPropsWithoutRef,
   type FC,
@@ -10,10 +8,10 @@ import {
 } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { useTheme } from '../../hooks/client/useTheme'
 import { useObjectAttributes } from '../../hooks/useObjectAttributes'
-import { Stack } from '../Layout'
 import { Text } from '../Text'
+
+import { ItemWrapper } from './client'
 
 type ObjectTermType = {
   text: ReactNode
@@ -54,7 +52,6 @@ export const DefinitionListItem: FC<Props> = ({
   fullWidth,
   className,
 }) => {
-  const theme = useTheme()
   const term = useObjectAttributes<ReactNode | ObjectTermType, ObjectTermType>(
     orgTerm,
     termObjectConverter,
@@ -68,27 +65,17 @@ export const DefinitionListItem: FC<Props> = ({
       term: cs.term(),
       description: cs.description(),
     }
-  }, [className, fullWidth])
+  }, [fullWidth, className])
 
   return (
-    <Stack
-      gap={0.25}
-      className={classNames.wrapper}
-      style={{
-        flexBasis:
-          // fullWidth の方が強い
-          !fullWidth && maxColumns
-            ? `calc((100% - ${theme.spacingByChar(1.5)} * ${maxColumns - 1}) / ${maxColumns})`
-            : undefined,
-      }}
-    >
+    <ItemWrapper maxColumns={maxColumns} fullWidth={fullWidth} className={classNames.wrapper}>
       <DefinitionTerm styleType={term.styleType} className={classNames.term}>
         {term.text}
       </DefinitionTerm>
       <Text as="dd" size="M" color="TEXT_BLACK" leading="NORMAL" className={classNames.description}>
         {children}
       </Text>
-    </Stack>
+    </ItemWrapper>
   )
 }
