+++ packages/smarthr-ui/src/components/DefinitionList/client/index.ts
@@ -1,0 +1,1 @@
+export { ItemWrapper } from './ItemWrapper'
